apps/server/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
psycopg2-binary
